[WPT/common/security-features] Merge generate.py (1/2)
This CL introduces ReferrerPolicy/MixedContestConfig classes in generate.py,
that encapsulates test-specific things, e.g.:
- Test description (moved from test_description.template)
- File paths (moved from common_paths.py etc.)
- The logic for generating <meta>/http headers (moved from
  `generate_selection()` in generate.py).

After this CL, the most part of
{mixed-content,referrer-policy}/generic/tools/generate.py are the same,
except for *Config classes.
This CL leaves the common parts in place in generate.py only for ease of
codereview, and a subsequent CL will merge them to common/.

- Functions in common_paths.py (which are not test-dependent) are moved to
  common/security-features/tools/util.py.
- test.js.template is removed and instead generator.py generates the content
  using json.dumps().
- Pretty printing of test.debug.html.template are removed.
- Other templates are merged/moved to common/security-features/tools/template.

This CL doesn't modify the generated files, because
the diffs in generated files don't affect behavior:
comments, location of `<meta>`, addition of `<meta charset='utf-8'>`,
and orders of JSON keys.
I'm planning to larger changes to generated files later, and will
regenerate files after that.

Bug: 906850
Change-Id: Ib509ede41e160b5b03876fe162d1d62c54e13151