Always leave an event in the AudioParam timeline
Previously, when the last AudioParam event has been processed and is
now in the past, the event would be removed from the timeline.  This
was an optimization so that we can skip sample-accurate processing of
the timeline and just returned fixed values.  However, if a new event
is added after all events have been deleted, the starting point of the
event could be incorrect.  For example a setValueAtTime event in the
past followed by a linearRampToValueAtTime.

Therefore, always leave the very last event in the timeline to
establish the starting event if needed by future events.

Bug: 951863
Change-Id: Iae0b550a4b497b953a10e85e603ecebb59508a99