[Animation Worklet] Use document from script state instead of the animated target
By spec [1] the ctor of Worklet Animation should use the Document that is associated
with the window that is the current global object. However the code takes the Document
from the target of the first effect which is not spec-compliant. e.g. the target is an
iframe, In addition, there may be security risks behind it.

[1] https://drafts.css-houdini.org/css-animationworklet/#creating-worklet-animation

Bug:953466
Change-Id: I13ed9a44e5a77f359e4075f1bd147d7e7163804a